fix: iterator with const and multilist support

+++ include/FIX-Message/Tag.hpp
@@ -16,7 +16,7 @@ namespace fix
     };
 
     template<fix::TagName Name, bool IsOptional>
-    struct TagNo : Tag<Name, std::conditional_t<IsOptional, std::optional<int>, int>>
+    struct TagNo : Tag<Name, std::conditional_t<IsOptional, std::optional<uint16_t>, uint16_t>>
     {
         static constexpr const bool Optional = IsOptional;
     };
